split out ListItem components
