+++ src/list/ListAvatar.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import Avatar from '../avatar/Avatar';
+
+function ListAvatar(props) {
+  return <Avatar size={40} rounded {...props} />;
+}
+
+export default ListAvatar;
