+++ src/list/__tests__/ListItem.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Text } from 'react-native';
+import { Text, Switch } from 'react-native';
 import { shallow } from 'enzyme';
 import toJson from 'enzyme-to-json';
 import { create } from 'react-test-renderer';
@@ -8,6 +8,11 @@ import theme from '../../config/theme';
 import { ThemeProvider } from '../../config';
 
 import ThemedListItem, { ListItem } from '../ListItem';
+import ListAvatar from '../ListAvatar';
+import ListIcon from '../ListIcon';
+import ListTitle from '../ListTitle';
+import ListChevron from '../ListChevron';
+import ListInput from '../ListInput';
 
 describe('ListItem component', () => {
   it('should render without issues', () => {
@@ -19,11 +24,9 @@ describe('ListItem component', () => {
 
   it('should render with avatar', () => {
     const component = shallow(
-      <ListItem
-        theme={theme}
-        avatar={{ source: 'avatar_uri' }}
-        containerStyle={{ backgroundColor: 'peru' }}
-      />
+      <ListItem theme={theme} containerStyle={{ backgroundColor: 'peru' }}>
+        <ListAvatar source="avatar_uri" />
+      </ListItem>
     );
 
     expect(component.length).toBe(1);
@@ -32,33 +35,20 @@ describe('ListItem component', () => {
 
   it('should render with left icon', () => {
     const component = shallow(
-      <ListItem
-        theme={theme}
-        leftIcon={{
-          name: 'wifi',
-          type: 'font-awesome',
-          color: 'red',
-          size: 20,
-        }}
-      />
+      <ListItem theme={theme}>
+        <ListIcon name="wifi" type="font-awesome" color="red" size={20} />
+      </ListItem>
     );
 
     expect(component.length).toBe(1);
     expect(toJson(component)).toMatchSnapshot();
   });
 
-  it('should render with left icon component', () => {
+  it('should render with Text component', () => {
     const component = shallow(
-      <ListItem theme={theme} leftIcon={<Text>I'm left icon</Text>} />
-    );
-
-    expect(component.length).toBe(1);
-    expect(toJson(component)).toMatchSnapshot();
-  });
-
-  it('should render with right icon component', () => {
-    const component = shallow(
-      <ListItem theme={theme} rightIcon={<Text>I'm right icon</Text>} />
+      <ListItem theme={theme}>
+        <Text>Test element</Text>
+      </ListItem>
     );
 
     expect(component.length).toBe(1);
@@ -67,15 +57,16 @@ describe('ListItem component', () => {
 
   it('should render with title and subtitle', () => {
     const component = shallow(
-      <ListItem
-        theme={theme}
-        title="title test"
-        titleProps={{ numberOfLines: 5 }}
-        subtitle="title test"
-        rightTitle="title"
-        titleStyle={{ backgroundColor: 'peru' }}
-        subtitleStyle={{ backgroundColor: 'peru' }}
-      />
+      <ListItem theme={theme}>
+        <ListTitle
+          title="title test"
+          titleProps={{ numberOfLines: 5 }}
+          subtitle="title test"
+          titleStyle={{ backgroundColor: 'peru' }}
+          subtitleStyle={{ backgroundColor: 'peru' }}
+        />
+        <ListTitle right title="title" />
+      </ListItem>
     );
 
     expect(component.length).toBe(1);
@@ -84,7 +75,10 @@ describe('ListItem component', () => {
 
   it('should render with switch', () => {
     const component = shallow(
-      <ListItem theme={theme} bottomDivider chevron switch={{ value: true }} />
+      <ListItem theme={theme} bottomDivider>
+        <Switch value={true} />
+        <ListChevron />
+      </ListItem>
     );
 
     expect(component.length).toBe(1);
@@ -93,7 +87,9 @@ describe('ListItem component', () => {
 
   it('should render with input', () => {
     const component = shallow(
-      <ListItem theme={theme} input={{ placeholder: 'Enter Text' }} />
+      <ListItem theme={theme}>
+        <ListInput placeholder="Enter Text" />
+      </ListItem>
     );
 
     expect(component.length).toBe(1);
@@ -102,14 +98,16 @@ describe('ListItem component', () => {
 
   it('should apply values from theme', () => {
     const testTheme = {
-      ListItem: {
+      ListTitle: {
         title: 'List Title',
       },
     };
 
     const component = create(
       <ThemeProvider theme={testTheme}>
-        <ThemedListItem />
+        <ThemedListItem>
+          <ListTitle />
+        </ThemedListItem>
       </ThemeProvider>
     );
